wapp-backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
